tools/display-giphy/requirements.txt
obs-websocket-py
python-dotenv
requests